requirements.txt
boto3
botocore
typer
pyyaml
pandas
